Harden C++ config validation
Agent-Logs-Url: https://github.com/Nightmare12139/CBBA-Python/sessions/afe45dc3-ae57-4ccd-8129-1dde4e27b915

+++ cpp/include/CBBA.h
@@ -78,7 +78,7 @@ private:
     std::array<double, 2> space_limit_y_{};
     std::array<double, 2> space_limit_z_{};
 
-    std::vector<int> time_interval_list_;
+    std::vector<double> time_interval_list_;
 
     std::vector<int> agent_index_list_;
     std::vector<std::vector<int>> bundle_list_;
